adams comments

+++ sdk/eventhub/azure-eventhub/azure/eventhub/_consumer_client.py
@@ -78,7 +78,7 @@ class EventHubConsumerClient(ClientBase):
      In 'exponential' mode, retry policy will sleep for: `{backoff factor} * (2 ** ({number of total retries} - 1))`
      seconds. If the backoff_factor is 0.1, then the retry will sleep
      for [0.0s, 0.2s, 0.4s, ...] between retries. The default value is 0.8.
-    :keyword int retry_backoff_max: The maximum back off time. Default value is 120 seconds (2 minutes).
+    :keyword float retry_backoff_max: The maximum back off time. Default value is 120 seconds (2 minutes).
     :keyword retry_mode: Fixed or exponential delay between attempts, default is exponential.
     :paramtype retry_mode: ~azure.eventhub.RetryMode
     :keyword float idle_timeout: Timeout, in seconds, after which this client will close the underlying connection
@@ -234,7 +234,7 @@ class EventHubConsumerClient(ClientBase):
          In 'exponential' mode, retry policy will sleep for: `{backoff factor} * (2 ** ({number of total retries} - 1))`
          seconds. If the backoff_factor is 0.1, then the retry will sleep
          for [0.0s, 0.2s, 0.4s, ...] between retries. The default value is 0.8.
-        :keyword int retry_backoff_max: The maximum back off time. Default value is 120 seconds (2 minutes).
+        :keyword float retry_backoff_max: The maximum back off time. Default value is 120 seconds (2 minutes).
         :keyword retry_mode: Fixed or exponential delay between attempts, default is exponential.
         :paramtype retry_mode: ~azure.eventhub.RetryMode
         :keyword float idle_timeout: Timeout, in seconds, after which this client will close the underlying connection
